+++ packages/terra-framework-docs/src/terra-dev-site/test/worklist-data-grid/PinnedColumns.test.jsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState, useCallback, useEffect } from 'react';
 import WorklistDataGrid from 'terra-worklist-data-grid';
 
 const gridDataJSON = {
@@ -54,20 +54,56 @@ const gridDataJSON = {
 const PinnedColumns = () => {
   const rowHeaderIndex = 0;
   const { cols } = gridDataJSON;
-  const [rows, setRows] = useState([]);
+  const [rowData, setRowData] = useState([]);
 
   // Change row data for data grid to validate that pinned column border resizes correctly.
   useEffect(() => {
-    setRows(gridDataJSON.rows);
+    setRowData(gridDataJSON.rows);
   }, []);
+
+  const onCellSelect = useCallback((rowId, columnId) => {
+    if (rowId && columnId) {
+      const rowIndex = rowData.findIndex(row => row.id === rowId);
+      const columnIndex = cols.findIndex(column => column.id === columnId);
+
+      // Remove current selections
+      const newRowData = [...rowData];
+      for (let row = 0; row < rowData.length; row += 1) {
+        for (let cell = 0; cell < rowData[row].cells.length; cell += 1) {
+          const currentCell = rowData[row].cells[cell];
+          if (currentCell.isSelected && !(row === rowIndex && cell === columnIndex)) {
+            currentCell.isSelected = false;
+          }
+        }
+      }
+
+      // Toggle selection state of selected cell
+      newRowData[rowIndex].cells[columnIndex].isSelected = !rowData[rowIndex].cells[columnIndex].isSelected;
+      setRowData(newRowData);
+    }
+  }, [cols, rowData]);
+
+  const onClearSelectedCells = useCallback(() => {
+    // Remove current selections
+    const newRowData = [...rowData];
+    for (let row = 0; row < rowData.length; row += 1) {
+      for (let cell = 0; cell < rowData[row].cells.length; cell += 1) {
+        newRowData[row].cells[cell].isSelected = false;
+      }
+    }
+
+    setRowData(newRowData);
+  }, [rowData]);
 
   return (
     <WorklistDataGrid
       id="pinned-columns"
       pinnedColumns={cols.slice(0, 3)}
       overflowColumns={cols.slice(3)}
-      rows={rows}
+      rows={rowData}
       rowHeaderIndex={rowHeaderIndex}
+      onCellSelect={onCellSelect}
+      onClearSelectedCells={onClearSelectedCells}
       ariaLabel="Worklist Data Grid"
     />
   );